# Conflicts:
#   backend/src/main/java/com/web/curation/controller/PhotoController.java
#   backend/src/main/java/com/web/curation/service/MailService.java

+++ backend/src/main/java/com/web/curation/service/PhotoServiceimpl.java
@@ -57,6 +57,38 @@ public class PhotoServiceimpl implements PhotoService {
         return getPhotoDtos(listCommunity);
     }
 
+    @Override
+    public PhotoDto detailPhoto(int boardId) {
+        Community community = communityRepository.findByBoardId(boardId);
+        // 상세 조회 시 조회 수 증가
+        int cnt = community.getClick();
+        community.setClick(cnt+1);
+        communityRepository.save(community);
+
+        PhotoDto photoDto = new PhotoDto();
+
+        photoDto.setProfileImgPath(community.getUser().getProfileImg());
+
+        photoDto.setBoardId(community.getBoardId());
+        photoDto.setNickname(community.getUser().getNickname());
+        photoDto.setContent(community.getContent());
+        photoDto.setHashtag(community.getHashtag());
+        photoDto.setUploadDate(community.getUploadDate());
+
+        // 좋아요 수
+        int countLike = communityLikeRepository.countByCommunity(community);
+        photoDto.setLike(countLike);
+
+        // 조회수
+        photoDto.setClick(community.getClick());
+
+        CommunityFile communityFile = communityFileRepository.findByCommunity(community);
+        photoDto.setFileName(communityFile.getName());
+        photoDto.setFilePath(communityFile.getFilePath());
+
+        return photoDto;
+    }
+
     private List<PhotoDto> getPhotoDtos(List<Community> listCommunity) {
         List<PhotoDto> listPhoto = new ArrayList<>();
 
@@ -66,20 +98,17 @@ public class PhotoServiceimpl implements PhotoService {
             photoDto.setProfileImgPath(community.getUser().getProfileImg());
 
             photoDto.setBoardId(community.getBoardId());
-            photoDto.setTitle(community.getTitle());
             photoDto.setNickname(community.getUser().getNickname());
             photoDto.setContent(community.getContent());
             photoDto.setHashtag(community.getHashtag());
             photoDto.setUploadDate(community.getUploadDate());
             
             // 좋아요 수
-            Long countLike = communityLikeRepository.countByCommunity(community);
+            int countLike = communityLikeRepository.countByCommunity(community);
             photoDto.setLike(countLike);
             
             // 조회수
             photoDto.setClick(community.getClick());
-            // photo인지 talk인지
-            photoDto.setDType(community.getDType());
 
 
             CommunityFile communityFile = communityFileRepository.findByCommunity(community);
@@ -93,12 +122,11 @@ public class PhotoServiceimpl implements PhotoService {
     }
     @Transactional
     @Override
-    public boolean writePhoto(PhotoDto photoDto) {
-        LOGGER.info("[photo 게시글 등록] 게시글 제목 : {}", photoDto.getTitle());
+    public int writePhoto(PhotoDto photoDto) {
+        LOGGER.info("[photo 게시글 등록] 게시글 ");
         // community 저장
         Community community = new Community();
 
-        community.setTitle(photoDto.getTitle());
         community.setContent(photoDto.getContent());
         community.setHashtag(photoDto.getHashtag());
         User user = userRepository.getByNickname(photoDto.getNickname());
@@ -106,7 +134,6 @@ public class PhotoServiceimpl implements PhotoService {
         community.setUser(user);
         community.setUploadDate(LocalDateTime.now());
         community.setClick(0);
-        community.setDType(2);
 
         Community savedCommunity = communityRepository.save(community);
 
@@ -120,12 +147,15 @@ public class PhotoServiceimpl implements PhotoService {
         communityFileRepository.save(communityFile);
 
         LOGGER.info("[getSignUpResult] userEntity 값이 들어왔는지 확인 후 결과값 주입");
+
+
         if (savedCommunity != null) {
             LOGGER.info("photo 게시글 저장 완료");
-            return true;
+            int boardId = savedCommunity.getBoardId();
+            return boardId;
         } else {
             LOGGER.info("photo 게시글 저장 실패");
-            return false;
+            return 0;
         }
     }
 
@@ -134,7 +164,6 @@ public class PhotoServiceimpl implements PhotoService {
 
         Community community = communityRepository.findByBoardId(photoDto.getBoardId());
 
-        community.setTitle(photoDto.getTitle());
         community.setContent(photoDto.getContent());
         community.setHashtag(photoDto.getHashtag());
 
@@ -149,8 +178,7 @@ public class PhotoServiceimpl implements PhotoService {
 
         communityFileRepository.save(communityFile);
 
-        if(community.getTitle().equals(photoDto.getTitle()) && community.getContent().equals(photoDto.getContent())
-        && community.getHashtag().equals(photoDto.getHashtag())){
+        if(community.getContent().equals(photoDto.getContent()) && community.getHashtag().equals(photoDto.getHashtag())){
             return true;
         }
         return false;
@@ -172,10 +200,12 @@ public class PhotoServiceimpl implements PhotoService {
     }
 
     @Override
-    public boolean pushLike(int boardId, String email) {
+    public int pushLike(int boardId, String email) {
         Community community = communityRepository.findByBoardId(boardId);
-
         User user = userRepository.getByEmail(email);
+
+        // 좋아요를 두번 누르면 fail
+        if(communityLikeRepository.existsByUserAndCommunity(user, community)) return -1;
 
         CommunityLike communityLike = new CommunityLike();
         communityLike.setCommunity(community);
@@ -183,29 +213,36 @@ public class PhotoServiceimpl implements PhotoService {
 
         communityLikeRepository.save(communityLike);
 
-        CommunityLike savedCommunityLike = communityLikeRepository.findByLikeId(communityLike.getLikeId());
+//        CommunityLike savedCommunityLike = communityLikeRepository.findByLikeId(communityLike.getLikeId());
+        // 좋아요 수
+        int countLike = communityLikeRepository.countByCommunity(community);
 
-        if (savedCommunityLike != null) {
+        if (communityLikeRepository.existsByLikeId(communityLike.getLikeId())) {
             LOGGER.info("push like 완료");
-            return true;
+            return countLike;
         } else {
             LOGGER.info("push like 실패");
-            return false;
+            return -1;
         }
     }
 
     @Override
-    public boolean cancelLike(int boardId, String email) {
+    public int cancelLike(int boardId, String email) {
 
         Community community = communityRepository.findByBoardId(boardId);
         User user = userRepository.getByEmail(email);
 
         CommunityLike communityLike = communityLikeRepository.findByCommunityAndUser(community, user);
+        // 좋아요 수
+        int countLike = communityLikeRepository.countByCommunity(community);
+
+        // 좋아요 수가 1이면 fail
+        if(countLike == 0) return -1;
 
         communityLikeRepository.delete(communityLike);
 
-        if(!communityLikeRepository.existsByLikeId(communityLike.getLikeId())) return true;
-        return false;
+        if(!communityLikeRepository.existsByLikeId(communityLike.getLikeId())) return countLike-1;
+        return -1;
     }
 
 }
